# fix(core): include archived tasks in health calculation to prevent drops