--- a/packages/core/src/adapters/metrics_adapter/index.ts
+++ b/packages/core/src/adapters/metrics_adapter/index.ts
@@ -126,7 +126,6 @@ export class MetricsAdapter implements IMetricsAdapter {
     this.cycleStore = dependencies.cycleStore;
     this.feedbackStore = dependencies.feedbackStore; // Graceful degradation
     this.executionStore = dependencies.executionStore; // Graceful degradation
-    this.changelogStore = dependencies.changelogStore; // Graceful degradation
     this.actorStore = dependencies.actorStore; // Graceful degradation
     this.platformApi = dependencies.platformApi; // Graceful degradation
   }
@@ -154,8 +153,6 @@ export class MetricsAdapter implements IMetricsAdapter {
 
     // Calculate Tier 1 metrics
     const health = this.calculateHealth(tasks);
-    const backlogDistribution = this.calculateBacklogDistribution(tasks);
-    const tasksCreatedToday = this.calculateTasksCreatedToday(tasks);
 
     // Count blocked and stale tasks
     const blockedTasks = tasks.filter(task => task.status === 'paused').length;
@@ -437,20 +434,25 @@ export class MetricsAdapter implements IMetricsAdapter {
     }
 
     try {
-      // Count tasks by status
-      const activeTasks = tasks.filter(task => task.status === 'active').length;
-      const doneTasks = tasks.filter(task => task.status === 'done').length;
-      const readyTasks = tasks.filter(task => task.status === 'ready').length;
-      const reviewTasks = tasks.filter(task => task.status === 'review').length;
-      const pausedTasks = tasks.filter(task => task.status === 'paused').length;
-      const draftTasks = tasks.filter(task => task.status === 'draft').length;
+      // Optimized: Single pass through tasks using reduce (performance improvement)
+      const statusCounts = tasks.reduce((acc, task) => {
+        acc[task.status] = (acc[task.status] || 0) + 1;
+        return acc;
+      }, {} as Record<string, number>);
 
       // Calculate health based on workflow progress and blockers
-      // Healthy tasks: done (100%), active (80%), ready (60%), review (40%)
+      // Healthy tasks: done (100%), archived (100%), active (80%), ready (60%), review (40%)
       // Neutral tasks: draft (20%)
       // Problematic tasks: paused (0%)
-      const healthyScore = (doneTasks * 100) + (activeTasks * 80) + (readyTasks * 60) + (reviewTasks * 40) + (draftTasks * 20) + (pausedTasks * 0);
-      
+      const healthyScore =
+        (statusCounts['done'] || 0) * 100 +
+        (statusCounts['archived'] || 0) * 100 +  // FIXED: Include archived tasks to prevent health drops
+        (statusCounts['active'] || 0) * 80 +
+        (statusCounts['ready'] || 0) * 60 +
+        (statusCounts['review'] || 0) * 40 +
+        (statusCounts['draft'] || 0) * 20 +
+        (statusCounts['paused'] || 0) * 0;
+
       const maxPossibleScore = tasks.length * 100;
 
       if (maxPossibleScore === 0) {
@@ -632,15 +634,13 @@ export class MetricsAdapter implements IMetricsAdapter {
       // Get unique agent IDs from recent executions
       const activeAgentIds = new Set<string>();
 
-      for (const execution of recentExecutions) {
+      if (recentExecutions.length > 0) {
         // Cross-reference executions with agent actors to find active agents
         const agentActors = actors.filter(actor => actor.type === 'agent');
         for (const agent of agentActors) {
           // In a complete implementation, we would track execution authorship
           // For now, if there are recent executions and agent actors, count them as active
-          if (recentExecutions.length > 0) {
-            activeAgentIds.add(agent.id);
-          }
+          activeAgentIds.add(agent.id);
         }
       }
 
@@ -704,42 +704,42 @@ export class MetricsAdapter implements IMetricsAdapter {
   /**
    * [EARS-15] Throws NotImplementedError for Tier 3 functions.
    */
-  calculateQuality(tasks: TaskRecord[]): number {
+  calculateQuality(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculateReworkRate(tasks: TaskRecord[]): number {
+  calculateReworkRate(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculateCompletionRate(tasks: TaskRecord[]): number {
+  calculateCompletionRate(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculateAuditScoreDistribution(tasks: TaskRecord[]): Record<string, number> {
+  calculateAuditScoreDistribution(_tasks: TaskRecord[]): Record<string, number> {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculateEpicPromotionRate(tasks: TaskRecord[]): number {
+  calculateEpicPromotionRate(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculateTaskRefinementRate(tasks: TaskRecord[]): number {
+  calculateTaskRefinementRate(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculatePlanningAccuracy(tasks: TaskRecord[]): number {
+  calculatePlanningAccuracy(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
-  calculateDependencyDiscoveryRate(tasks: TaskRecord[]): number {
+  calculateDependencyDiscoveryRate(_tasks: TaskRecord[]): number {
     throw new Error('NotImplementedError: Tier 3 metrics not implemented yet');
   }
 
   /**
    * [EARS-16] Returns null for Premium metrics without Platform API.
    */
-  calculateCostBurnRate(consumption: TokenConsumption[]): number {
+  calculateCostBurnRate(_consumption: TokenConsumption[]): number {
     if (!this.platformApi) {
       console.warn('Platform API not available for premium metrics');
       return 0;
@@ -747,7 +747,7 @@ export class MetricsAdapter implements IMetricsAdapter {
     throw new Error('NotImplementedError: Tier 4 premium metrics not implemented yet');
   }
 
-  calculateTokenConsumption(consumption: TokenConsumption[]): number {
+  calculateTokenConsumption(_consumption: TokenConsumption[]): number {
     if (!this.platformApi) {
       console.warn('Platform API not available for premium metrics');
       return 0;
@@ -755,7 +755,7 @@ export class MetricsAdapter implements IMetricsAdapter {
     throw new Error('NotImplementedError: Tier 4 premium metrics not implemented yet');
   }
 
-  calculateTokenConsumptionByAgent(consumption: TokenConsumption[]): Record<string, number> {
+  calculateTokenConsumptionByAgent(_consumption: TokenConsumption[]): Record<string, number> {
     if (!this.platformApi) {
       console.warn('Platform API not available for premium metrics');
       return {};
@@ -763,7 +763,7 @@ export class MetricsAdapter implements IMetricsAdapter {
     throw new Error('NotImplementedError: Tier 4 premium metrics not implemented yet');
   }
 
-  calculateAiAccuracyRate(tasks: TaskRecord[], feedback: FeedbackRecord[]): number {
+  calculateAiAccuracyRate(_tasks: TaskRecord[], _feedback: FeedbackRecord[]): number {
     if (!this.platformApi) {
       console.warn('Platform API not available for premium metrics');
       return 0;
@@ -771,7 +771,7 @@ export class MetricsAdapter implements IMetricsAdapter {
     throw new Error('NotImplementedError: Tier 4 premium metrics not implemented yet');
   }
 
-  calculateAgentExecutionTime(executions: ExecutionRecord[]): number {
+  calculateAgentExecutionTime(_executions: ExecutionRecord[]): number {
     if (!this.platformApi) {
       console.warn('Platform API not available for premium metrics');
       return 0;