## 🐛 Critical Bug Fixed: Health Calculation Drops

**Problem:** Health score was dropping dramatically (90% → 2%) when archiving completed tasks because archived tasks were excluded from the calculation.

## 🎯 Task Reference
**GitGovernance Task:** `1758539111-task-fix-health-calculation-to-include-archived-tasks`

## ✨ Solution Implemented

### 🔧 **Bug Fix: Include Archived Tasks**
- ✅ Added archived tasks with **100% weight** (same as done tasks)
- ✅ Prevents health drops during healthy archiving operations
- ✅ Follows blueprint specification v2.1

### ⚡ **Performance Optimization**
- 🚀 Optimized from **6 separate filters** to **1 reduce operation**
- 📈 Improved from **O(6n)** to **O(n)** complexity
- ⏱️ Maintains **<100ms** performance target (EARS-4)

### 🧪 **Enhanced Test Coverage**
- 🔄 Updated **EARS-8** test to include archived tasks
- 🛡️ Added **EARS-8.1** regression test to prevent future drops
- ✅ All **33/33 tests** passing

## 📁 Files Changed
```
📝 packages/core/src/adapters/metrics_adapter/index.ts
🧪 packages/core/src/adapters/metrics_adapter/metrics_adapter.test.ts
```

## 🔍 Code Preview
```typescript
// 🔥 NEW: Optimized single-pass calculation with archived fix
const statusCounts = tasks.reduce((acc, task) => {
  acc[task.status] = (acc[task.status] || 0) + 1;
  return acc;
}, {} as Record<string, number>);

const healthyScore = 
  (statusCounts['done'] || 0) * 100 +
  (statusCounts['archived'] || 0) * 100 +  // 🎯 FIXED: Include archived!
  (statusCounts['active'] || 0) * 80 +
  // ... rest of calculation
```

## ✅ Validation Checklist
- 🟢 **TypeScript**: No errors in packages/core
- 🟢 **Tests**: 33/33 passing including regression test
- 🟢 **Performance**: Maintains <100ms target
- 🟢 **Blueprint**: Implements documented v2.1 formula
- 🟢 **EARS**: Updated EARS-8, added EARS-8.1

## 📊 Impact
| Before | After |
|--------|-------|
| 🔴 Health drops 90% → 2% when archiving | 🟢 Health stable 100% → 100% when archiving |
| 🐌 6 separate filter operations | ⚡ 1 optimized reduce operation |
| ❌ Penalizes healthy archiving | ✅ Rewards completed work |

**Result:** Archiving completed work (healthy operation) no longer penalizes project health! 🎉